+++ src/logic/createFormControl.ts
@@ -30,6 +30,7 @@ import {
   UseFormRegister,
   UseFormRegisterReturn,
   UseFormReset,
+  UseFormResetField,
   UseFormReturn,
   UseFormSetError,
   UseFormSetFocus,
@@ -67,11 +68,13 @@ import omit from '../utils/omit';
 import unset from '../utils/unset';
 
 import focusFieldBy from './focusFieldBy';
+import { generateWatchOutput } from './generateWatchOutput';
 import getFieldValue from './getFieldValue';
 import getFieldValueAs from './getFieldValueAs';
 import getResolverOptions from './getResolverOptions';
 import hasValidation from './hasValidation';
 import isNameInFieldArray from './isNameInFieldArray';
+import isWatched from './isWatched';
 import schemaErrorLookup from './schemaErrorLookup';
 import setFieldArrayDirtyFields from './setFieldArrayDirtyFields';
 import skipValidation from './skipValidation';
@@ -137,7 +140,6 @@ export function createFormControl<
   };
   const _subjects: Subjects<TFieldValues> = {
     watch: createSubject(),
-    control: createSubject(),
     array: createSubject(),
     state: createSubject(),
   };
@@ -153,15 +155,6 @@ export function createFormControl<
       clearTimeout(timer);
       timer = window.setTimeout(() => callback(...args), wait);
     };
-
-  const isFieldWatched = (
-    name: FieldPath<TFieldValues>,
-    isBlurEvent?: boolean,
-  ) =>
-    !isBlurEvent &&
-    (_names.watchAll ||
-      _names.watch.has(name) ||
-      _names.watch.has((name.match(/\w+/) || [])[0]));
 
   const _updateValid = async (shouldSkipRender?: boolean) => {
     let isValid = false;
@@ -496,19 +489,7 @@ export function createFormControl<
         : defaultValue),
     };
 
-    if (names) {
-      const result = convertToArrayPayload(names).map(
-        (fieldName) => (
-          isGlobal && _names.watch.add(fieldName as InternalFieldName),
-          get(fieldValues, fieldName as InternalFieldName)
-        ),
-      );
-
-      return Array.isArray(names) ? result : result[0];
-    }
-
-    isGlobal && (_names.watchAll = true);
-    return fieldValues;
+    return generateWatchOutput(names, _names, fieldValues, isGlobal);
   };
 
   const _getFieldArray = (name: InternalFieldName) =>
@@ -522,7 +503,6 @@ export function createFormControl<
     name: InternalFieldName,
     value: SetFieldValue<TFieldValues>,
     options: SetValueConfig = {},
-    shouldRender?: boolean,
   ) => {
     const field: Field = get(_fields, name);
     let fieldValue: unknown = value;
@@ -568,12 +548,6 @@ export function createFormControl<
         } else {
           fieldReference.ref.value = fieldValue;
         }
-
-        shouldRender &&
-          _subjects.control.next({
-            values: _formValues,
-            name,
-          });
       }
     }
 
@@ -598,7 +572,7 @@ export function createFormControl<
         (field && !field._f)) &&
       !isDateObject(fieldValue)
         ? setValues(fieldName, fieldValue, options)
-        : setFieldValue(fieldName, fieldValue, options, true);
+        : setFieldValue(fieldName, fieldValue, options);
     }
   };
 
@@ -633,10 +607,10 @@ export function createFormControl<
     } else {
       field && !field._f && !isNullOrUndefined(value)
         ? setValues(name, value, options)
-        : setFieldValue(name, value, options, true);
+        : setFieldValue(name, value, options);
     }
 
-    isFieldWatched(name) && _subjects.state.next({});
+    isWatched(name, _names) && _subjects.state.next({});
     _subjects.watch.next({
       name,
     });
@@ -664,7 +638,7 @@ export function createFormControl<
           validationModeAfterSubmit,
           validationModeBeforeSubmit,
         );
-      const isWatched = isFieldWatched(name, isBlurEvent);
+      const watched = isWatched(name, _names, isBlurEvent);
 
       if (isBlurEvent) {
         field._f.onBlur && field._f.onBlur(event);
@@ -681,7 +655,7 @@ export function createFormControl<
         false,
       );
 
-      const shouldRender = !isEmptyObject(fieldState) || isWatched;
+      const shouldRender = !isEmptyObject(fieldState) || watched;
 
       !isBlurEvent &&
         _subjects.watch.next({
@@ -692,11 +666,11 @@ export function createFormControl<
       if (shouldSkipValidation) {
         return (
           shouldRender &&
-          _subjects.state.next({ name, ...(isWatched ? {} : fieldState) })
+          _subjects.state.next({ name, ...(watched ? {} : fieldState) })
         );
       }
 
-      !isBlurEvent && isWatched && _subjects.state.next({});
+      !isBlurEvent && watched && _subjects.state.next({});
 
       validateFields[name] = validateFields[name] ? +1 : 1;
 
@@ -770,7 +744,7 @@ export function createFormControl<
           }),
         )
       ).every(Boolean);
-      _updateValid();
+      !(!validationResult && !_formState.isValid) && _updateValid();
     } else {
       validationResult = isValid = await executeBuildInValidation(_fields);
     }
@@ -910,7 +884,9 @@ export function createFormControl<
     });
     _names.mount.add(name);
 
-    !isUndefined(options.value) && set(_formValues, name, options.value);
+    !isUndefined(options.value) &&
+      !options.disabled &&
+      set(_formValues, name, get(_formValues, name, options.value));
 
     field
       ? isBoolean(options.disabled) &&
@@ -1001,7 +977,9 @@ export function createFormControl<
         e.persist && e.persist();
       }
       let hasNoPromiseError = true;
-      let fieldValues: any = { ..._formValues };
+      let fieldValues: any = _options.shouldUnregister
+        ? cloneObject(_formValues)
+        : { ..._formValues };
 
       _subjects.state.next({
         isSubmitting: true,
@@ -1050,13 +1028,42 @@ export function createFormControl<
       }
     };
 
+  const resetField: UseFormResetField<TFieldValues> = (name, options = {}) => {
+    if (isUndefined(options.defaultValue)) {
+      setValue(name, get(_defaultValues, name));
+    } else {
+      setValue(name, options.defaultValue);
+      set(_defaultValues, name, options.defaultValue);
+    }
+
+    if (!options.keepTouched) {
+      unset(_formState.touchedFields, name);
+    }
+
+    if (!options.keepDirty) {
+      unset(_formState.dirtyFields, name);
+      _formState.isDirty = options.defaultValue
+        ? _getDirty(name, get(_defaultValues, name))
+        : _getDirty();
+    }
+
+    if (!options.keepError) {
+      unset(_formState.errors, name);
+      _proxyFormState.isValid && _updateValid();
+    }
+
+    _subjects.state.next({ ..._formState });
+  };
+
   const reset: UseFormReset<TFieldValues> = (
     formValues,
     keepStateOptions = {},
   ) => {
-    const hasUpdatedFormValues = !isEmptyObject(formValues);
     const updatedValues = formValues || _defaultValues;
     const cloneUpdatedValues = cloneObject(updatedValues);
+    const values = !isEmptyObject(formValues)
+      ? cloneUpdatedValues
+      : _defaultValues;
 
     if (!keepStateOptions.keepDefaultValues) {
       _defaultValues = updatedValues;
@@ -1087,14 +1094,12 @@ export function createFormControl<
         : cloneUpdatedValues;
       _fields = {};
 
-      _subjects.control.next({
-        values: hasUpdatedFormValues ? cloneUpdatedValues : _defaultValues,
+      _subjects.watch.next({
+        values,
       });
 
-      _subjects.watch.next({});
-
       _subjects.array.next({
-        values: cloneUpdatedValues,
+        values,
       });
     }
 
@@ -1216,6 +1221,7 @@ export function createFormControl<
     setValue,
     getValues,
     reset,
+    resetField,
     clearErrors,
     unregister,
     setError,