+++ src/useController.ts
@@ -6,7 +6,7 @@ import get from './utils/get';
 import { EVENTS } from './constants';
 import {
   Field,
-  FieldPathWithValue,
+  FieldPath,
   FieldValues,
   InternalFieldName,
   UseControllerProps,
@@ -14,60 +14,49 @@ import {
 } from './types';
 import { useFormContext } from './useFormContext';
 import { useFormState } from './useFormState';
-import { useSubscribe } from './useSubscribe';
+import { useWatch } from './useWatch';
 
 export function useController<
   TFieldValues extends FieldValues = FieldValues,
-  TResult = any,
-  TName extends FieldPathWithValue<TFieldValues, TResult> = FieldPathWithValue<
-    TFieldValues,
-    TResult
-  >,
+  TName extends FieldPath<TFieldValues> = FieldPath<TFieldValues>,
 >(
-  props: UseControllerProps<TFieldValues, TResult, TName>,
-): UseControllerReturn<TFieldValues, TResult, TName> {
+  props: UseControllerProps<TFieldValues, TName>,
+): UseControllerReturn<TFieldValues, TName> {
   const methods = useFormContext<TFieldValues>();
   const { name, control = methods.control, shouldUnregister } = props;
-  const [value, setInputStateValue] = React.useState(
-    get(
+  const isArrayField = isNameInFieldArray(control._names.array, name);
+  const value = useWatch({
+    control,
+    name,
+    defaultValue: get(
       control._formValues,
       name,
       get(control._defaultValues, name, props.defaultValue),
     ),
-  );
+    exact: !isArrayField,
+  });
   const formState = useFormState({
-    control: control || methods.control,
+    control,
     name,
   });
   const _name = React.useRef(name);
 
   _name.current = name;
 
-  useSubscribe({
-    subject: control._subjects.control,
-    callback: (data) =>
-      (!data.name || _name.current === data.name) &&
-      setInputStateValue(get(data.values, _name.current)),
-    skipEarlySubscription: true,
-  });
-
   const registerProps = control.register(name, {
     ...props.rules,
     value,
   });
 
-  const updateMounted = React.useCallback(
-    (name: InternalFieldName, value: boolean) => {
+  React.useEffect(() => {
+    const updateMounted = (name: InternalFieldName, value: boolean) => {
       const field: Field = get(control._fields, name);
 
       if (field) {
         field._f.mount = value;
       }
-    },
-    [control],
-  );
+    };
 
-  React.useEffect(() => {
     updateMounted(name, true);
 
     return () => {
@@ -75,26 +64,23 @@ export function useController<
         control._options.shouldUnregister || shouldUnregister;
 
       if (
-        isNameInFieldArray(control._names.array, name)
+        isArrayField
           ? _shouldUnregisterField && !control._stateFlags.action
           : _shouldUnregisterField
       ) {
-        control.unregister(name);
+        control.unregister(name, { keepDefaultValue: true });
       } else {
         updateMounted(name, false);
       }
     };
-  }, [name, control, shouldUnregister, updateMounted]);
+  }, [name, control, isArrayField, shouldUnregister]);
 
   return {
     field: {
-      onChange: (event: any) => {
-        const value = getControllerValue(event);
-        setInputStateValue(value);
-
+      onChange: (event) => {
         registerProps.onChange({
           target: {
-            value,
+            value: getControllerValue(event),
             name: name as InternalFieldName,
           },
           type: EVENTS.CHANGE,